requirements.txt
aesthetic_predictor_v2_5==2024.12.18.1
decord==0.6.0
diffusers==0.32.2
einops==0.8.1
gradio==5.33.1
huggingface_hub==0.28.1
hydra-core==1.3.2
iopath==0.1.10
modelscope==1.23.2
omegaconf==2.3.0
opencv_python==4.11.0.86
optimum_quanto==0.2.6
pandas==1.5.3
peft==0.14.0
Pillow==11.2.1
preprocessing==0.1.13
pycocotools==2.0.8
pyfacer==0.0.5
PyYAML==6.0.2
Requests==2.32.4
safetensors==0.5.3
supervision==0.25.1
timm==1.0.15
torchao==0.11.0
tqdm==4.67.1
webdataset==0.2.111
transformers==4.49.0
sentencepiece
protobuf
